Extract `ClientStorage` mock/helper
